[fire-drill-2.2.1] Incompatible framework upgrade
This PR introduces an incompatible change to the framework (a new
parameter for a public function), to test incompatible upgrades.

It is intended to be used to build binaries that are used temporarily
to confirm that a protocol version upgrade involving an incompatible
framework upgrade will not go through.

Test Plan:

Fire drills